// Client/src/app/core/models/afmeEmployee.ts
export class AfmeEmployee {
    marketName: string;
    mobileNumber: string;
    country: string;
    departmentName: string;
    emailAddress: string;
    id: number;
    firstName: string;
    globalID: string;
    jobTitle: string;
    lastName: string;
    employeeDomain: string;
    managerEmailAddress: string;
    managerName: string;
    managerNetworkID: string;
    managerPeopleSoft: string;
    managerDomain: string;
    managerOfManagerEmailAddress: string;
    managerOfManagerName: string;
    managerOfManagerNetworkID: string;
    managerOfManagerPeopleSoft: string;
    middleName: string;
    fullName: string;
    networkLoginId: string;
    peopleSoftId: number;
    userName: string;
    dataOfHiring: string;
    displayName: string;
    businessUnit: string;
    isEmployee: string;
    companyCode: string;
    companyName: string;
    ACQ_DIV_FLAG: string;
    ELIG_CONFIG6: string;
    vendorName: string;
}